20190625 Fortify Updates to fix Critical and High Fortify issues
* Fixed Null Dereference in CSSHPAUserDefinedDes and PIAPEB_WrapTest
* Fixed Unreleased Resource Streams in BasicWriterTest,
CSSHPAUserDefinedDES, CSSHPA_CreateTest, FileComparer, Jpeg2000Test,
JpegTest, Nitf21HeaderTest, SensorLookupTest, and SimpleLookupTest
* Fixed XML External Entity Injection in RasterProductFormatUtilities,
SensorLookup, and TreParser

+++ deswrap/src/main/java/org/codice/imaging/nitf/deswrap/CSSHPAUserDefinedDES.java
@@ -121,7 +121,9 @@ class CSSHPAUserDefinedDES implements UserDefinedDataExtensionSegment {
             } catch (IOException ex) {
                 LOG.warn("Could not generate CSSHPA.", ex);
             } finally {
-                tempFile.delete();
+            	if (tempFile != null) {
+            		tempFile.delete();
+            	}
                 try {
                     desData.close();
                 } catch (IOException ex) {
@@ -187,7 +189,11 @@ class CSSHPAUserDefinedDES implements UserDefinedDataExtensionSegment {
                 shapeClassName = "MULTIPATCH";
                 break;
             default:
+            	shp.close();
                 throw new IOException(String.format("Error reading shapefile, unknown shapefile geometry class: %d", shapefileClass));
+        }
+        if (shp != null) {
+        	shp.close();
         }
         return padStringToLength(shapeClassName, CSSHPAConstants.SHAPE_CLASS_LENGTH);
     }
